refactor: move action class to pyBrematic root
This change should still be backwards compatible.
