+++ pyBrematic/__init__.py
@@ -1,5 +1,6 @@
 from .devices.device import Device
 from .gateways.brennenstuhl_gateway import BrennenstuhlGateway
 from .gateways.gateway import Gateway
+from .action import Action
 
-__all__ = ('Device', 'Gateway', 'BrennenstuhlGateway')
+__all__ = ["Device", "Gateway", "BrennenstuhlGateway", "Action"]
