+++ pyBrematic/devices/tests/device_test.py
@@ -2,7 +2,8 @@
 
 import unittest
 
-from pyBrematic.devices import Device, Action
+from pyBrematic.action import Action
+from pyBrematic.devices import Device
 from pyBrematic.gateways import BrennenstuhlGateway
